Conflicts:
	.gitignore
	CMakeLists.txt
	CMakeModules/FindSvnClient.cmake
	Makefile.am
	configure.ac
	projects/VC90/.gitignore
	projects/VC90/FlightGear.sln
	src/AIModel/AIBallistic.cxx
	src/Instrumentation/vertical_speed_indicator.cxx
	src/Main/CMakeLists.txt
	src/Main/CameraGroup.cxx
	src/Main/renderer.cxx
	src/Network/Makefile.am
	src/Viewer/FGEventHandler.cxx

Conflicts resolved by picking versions from release/2.10.0

+++ package/mac/build-mac-nightly-dmg.rb
@@ -3,7 +3,7 @@
 require 'ERB'
 
 $osgLibs = ['osgFX', 'osgParticle', 'osg', 'osgGA', 'osgText', 'osgUtil', 'osgSim', 'osgViewer', 'osgDB']
-$osgPlugins = ['ac', 'osg', 'freetype', 'qt', 'imageio', 'rgb', 'txf', 'mdl', '3ds']
+$osgPlugins = ['ac', 'osg', 'freetype', 'imageio', 'rgb', 'txf', 'mdl', '3ds', 'dds']
 
 def runOsgVersion(option)
   env = "export DYLD_LIBRARY_PATH=#{Dir.pwd}/dist/lib"
@@ -15,9 +15,14 @@ osgVersion = runOsgVersion('version-number')
 $osgSoVersion=runOsgVersion('so-number')
 $openThreadsSoVersion=runOsgVersion('openthreads-soversion-number')
 
+$codeSignIdentity = ENV['FG_CODESIGN_IDENTITY']
+puts "Code signing identity is #{$codeSignIdentity}"
+
 puts "osgVersion=#{osgVersion}, so-number=#{$osgSoVersion}"
 
-$alutSourcePath='/Library/Frameworks/ALUT.framework'
+$svnLibs = ['svn_client', 'svn_wc', 'svn_delta', 'svn_diff', 'svn_ra', 
+  'svn_ra_local', 'svn_repos', 'svn_fs', 'svn_fs_fs', 'svn_fs_util',
+  'svn_ra_svn', 'svn_subr', 'svn_ra_neon']
 
 def fix_install_names(object)
   #puts "fixing install names for #{object}"
@@ -31,16 +36,11 @@ def fix_install_names(object)
   oldName = "libOpenThreads.#{$openThreadsSoVersion}.dylib"
   newName= "@executable_path/../Frameworks/#{oldName}"
   `install_name_tool -change #{oldName} #{newName} #{object}`
-  
-  alutBundlePath = "@executable_path/../Frameworks/Alut.framework"
-  alutLib = "Versions/A/ALUT"
-  `install_name_tool -change #{$alutSourcePath}/#{alutLib} #{alutBundlePath}/#{alutLib} #{object}`
 end
 
-prefixDir=Dir.pwd + "/dist"
+$prefixDir=Dir.pwd + "/dist"
 dmgDir=Dir.pwd + "/image"
 srcDir=Dir.pwd + "/flightgear"
-
 
 puts "Erasing previous image dir"
 `rm -rf #{dmgDir}`
@@ -48,10 +48,34 @@ puts "Erasing previous image dir"
 bundle=dmgDir + "/FlightGear.app"
 contents=bundle + "/Contents"
 macosDir=contents + "/MacOS"
-frameworksDir=contents +"/Frameworks"
+$frameworksDir=contents +"/Frameworks"
 resourcesDir=contents+"/Resources"
 osgPluginsDir=contents+"/PlugIns/osgPlugins-#{osgVersion}"
 volName="\"FlightGear Nightly Build\""
+
+def fix_svn_install_names(object)
+  $svnLibs.each do |l|
+    fileName = "lib#{l}-1.0.dylib"
+    newName = "@executable_path/../Frameworks/#{fileName}"
+    `install_name_tool -change #{fileName} #{newName} #{object}`
+  end
+end
+
+def copy_svn_libs()
+  puts "Copying Subversion client libraries"
+  $svnLibs.each do |l|
+    libFile = "lib#{l}-1.0.dylib"
+    path = "#{$frameworksDir}/#{libFile}"
+    `cp #{$prefixDir}/lib/#{libFile} #{$frameworksDir}`
+    fix_svn_install_names(path)
+   # `install_name_tool -id #{libFile}  #{path}`    
+  end
+end
+
+def code_sign(path)
+  puts "Signing #{path}"
+  `codesign -s "#{$codeSignIdentity}" #{path}`
+end
 
 fgVersion = File.read("#{srcDir}/version").strip
 
@@ -59,37 +83,41 @@ dmgPath = Dir.pwd + "/fg_mac_nightly_#{fgVersion}.dmg"
 
 puts "Creating directory structure"
 `mkdir -p #{macosDir}`
-`mkdir -p #{frameworksDir}`
+`mkdir -p #{$frameworksDir}`
 `mkdir -p #{resourcesDir}`
 `mkdir -p #{osgPluginsDir}`
 
 puts "Copying binaries"
-bins = ['fgfs', 'terrasync', 'fgjs']
+bins = ['fgfs', 'fgjs', 'fgcom']
 bins.each do |b|
-  `cp #{prefixDir}/bin/#{b} #{macosDir}/#{b}`
-  fix_install_names("#{macosDir}/#{b}")
+  if !File.exist?("#{$prefixDir}/bin/#{b}")
+    next
+  end
+  
+  outPath = "#{macosDir}/#{b}"
+  `cp #{$prefixDir}/bin/#{b} #{outPath}`
+  fix_install_names(outPath)
+  fix_svn_install_names(outPath)
 end
 
 puts "copying libraries"
 $osgLibs.each do |l|
   libFile = "lib#{l}.#{$osgSoVersion}.dylib"
-  `cp #{prefixDir}/lib/#{libFile} #{frameworksDir}`
-  fix_install_names("#{frameworksDir}/#{libFile}")
+  `cp #{$prefixDir}/lib/#{libFile} #{$frameworksDir}`
+  fix_install_names("#{$frameworksDir}/#{libFile}")
 end
 
 # and not forgetting OpenThreads
 libFile = "libOpenThreads.#{$openThreadsSoVersion}.dylib"
-`cp #{prefixDir}/lib/#{libFile} #{frameworksDir}`
+`cp #{$prefixDir}/lib/#{libFile} #{$frameworksDir}`
 
 $osgPlugins.each do |p|
   pluginFile = "osgdb_#{p}.so"
-  `cp #{prefixDir}/lib/osgPlugins-#{osgVersion}/#{pluginFile} #{osgPluginsDir}`
+  `cp #{$prefixDir}/lib/osgPlugins-#{osgVersion}/#{pluginFile} #{osgPluginsDir}`
   fix_install_names("#{osgPluginsDir}/#{pluginFile}")
 end
 
-# custom ALUT
-# must copy frameworks using ditto
-`ditto #{$alutSourcePath} #{frameworksDir}/ALUT.framework`
+copy_svn_libs()
 
 # Info.plist
 template = File.read("#{srcDir}/package/mac/nightly.plist.in")
@@ -104,11 +132,26 @@ File.open("#{contents}/Info.plist", 'w') { |f|
 `cp #{srcDir}/COPYING #{dmgDir}`
 
 # Macflightgear launcher
-puts "Copying Macflightgear launcher files"
+if File.exist?("FlightGearOSX")
+  puts "Copying Macflightgear launcher files"
+  Dir.chdir "FlightGearOSX" do
+    `cp FlightGear #{macosDir}`
+    `rsync -a *.rb *.lproj *.sh *.tiff #{resourcesDir}`
+  end
+end
 
-Dir.chdir "macflightgear" do
-  `cp FlightGear #{macosDir}`
-  `rsync -a --exclude=\".svn\" *.rb *.lproj *.sh *.tiff #{resourcesDir}`
+if File.exist?("#{$prefixDir}/bin/fgcom-data")
+  puts "Copying FGCom data files"
+  `ditto #{$prefixDir}/bin/fgcom-data #{resourcesDir}/fgcom-data`
+end
+
+# code sign all executables in MacOS dir. Do this last since reource
+# changes will invalidate the signature!
+Dir.foreach(macosDir) do |b|
+    if b == '.' or b == '..' then
+        next
+    end
+  code_sign("#{macosDir}/#{b}")
 end
 
 puts "Creating DMG"
